Migrated the useLoginModal hook back to the Login.js component file so it stays in close proximity with the other Login component code.
