+++ client/src/components/Modal/utils.js
@@ -39,22 +39,3 @@ export const useOpenModal = ( key ) =>{
 		}),[key]);
 
 }
-
-export const useLoginModal = () =>{
-    const modalKey = "LOGIN_MODAL"
-    useModalRegistration(modalKey, {
-        key: modalKey,
-        component: () => (
-            <Box className="py-5">
-                <Heading renderAs="h2">Login</Heading>
-                <hr />
-                <LoginForm />
-                <p className="mt-3 has-text-grey is-size-7">
-                    Don't have an account yet?{" "}
-                    <a href="/register">Register</a>
-                </p>
-            </Box>
-        )
-    })
-	
-}
